Add OverloadedStrings to default-extensions
This software currently does not make use of `OverloadedStrings`.
Enabling the extension ensures that future changes will be compatible
with it, however.  (Modules that do not have the extension enabled and
use string literals that are ambiguous when the extension is enabled
fail to load in `ghci` when `ghci` is configured to use the extension.)