Refactor Makefile
MODE is now used directly, with values `cabal` or `stack`, instead of
defaulting to Stack and using Cabal when CABAL is set.  The default MODE
is set in the project configuration.

GHC_VERSION is now exposed.  It is set automatically based on the
version of `ghc` when not set.

A `--project-file` option is added to CABAL_ARGS if PROJECT_FILE is set,
as before.  When not set, project file `cabal-$(GHC_VERSION).project` is
used if it exists.

CABAL_TEST_GHC_VERSIONS is added, and the `test-all` command uses
recursive make to run tests for these versions.