Add test-all.sh
This script parses the `.cabal` file to determine which GHC versions to
test.  The `cabal` and `stack` commands are used to test all versions
using Cabal and Stack, respectively.  The `github` command outputs the
versions to set the GitHub Actions matrix.

The list of GHC versions to test is now only configured in the `.cabal`
file; there is now no need to update the versions in the `Makefile` or
GitHub Actions configuration.  Note, however, that the Cabal version
tests are still configured in the GitHub Actions configuration.