Refactor GitHub Actions CI (WIP)
In preparation for managing multiple (related) packages within this
repository, I am refactoring the GitHub Actions CI.  Once I get things
working with manual configuration, I plan on writing a script to
generate the configuration, making it easier to maintain.  This commit
is the very first test, so there are likely issues.