Audit type argument order
This commit adds explicit `forall` to all function type signatures that
have more than one type argument.  The goal is to make it easier/faster
to understand the order of type arguments when using `TypeApplications`.

Note that this is not done for methods, which include type arguments
that are bound at the type class level.

This commit does not change the order of any type arguments.