Use latest Stack in CI tests
A month ago, `stack-version: latest` did not install the latest version
of Stack, causing GHC 9.10.1 tests to fail.  I figured that `latest` may
install the version that is "recommended" by GHCup.

Today, I see in other CI results that `latest` installed the latest
version (2.15.7) of Stack.  Checking GHCup, the "recommended" version
has not changed.

This commit changes the configuration back to `latest`.  Hopefully it
works and is consistent with what I see elsewhere!